Move RunConfigSettings into it's own class (net.fabricmc.loom.configuration.RunConfigSettings)
